pyproject.toml
[project]
name = "agentic-employees"
version = "0.1.0"
description = "Event-driven agentic employees with RL-based self-improvement"
requires-python = ">=3.10"
dependencies = [
    "nats-py>=2.9.0",
    "pydantic>=2.0",
    "ollama>=0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]
training = [
    "torch>=2.0",
    "transformers>=4.40",
    "peft>=0.11",
]
inference = [
    "openai>=1.0",
    "httpx>=0.27",
]
vllm = [
    "vllm>=0.8.1",
]
openrlhf = [
    "openrlhf[vllm]>=0.9.0",
]
bridge = [
    "aiohttp>=3.9",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.ruff]
target-version = "py310"
